updated cabal file and added a bunch of new functions for getting data on pokemon.
The functions form TypeComparison have been moved over to Pokedex for
quicker testing in the repl, the will be moved back later when I have
decided on the exact interface etc.
PokeCache is now its own module exporting the pokedex. This solves the
issue of having the pokedex in the IO monad